requirements.txt
numpy==1.26.4
pandas
scipy
scikit-learn
omegaconf
hydra-core
tensorboard
transformers
